+++ src/main/java/com/aqanetics/agent/config/AqaConfigLoader.java
@@ -1,4 +1,4 @@
-package com.aqanetics;
+package com.aqanetics.agent.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
