+++ src/main/java/com/aqanetics/agent/core/dto/NewSuiteExecutionDto.java
@@ -1,4 +1,4 @@
-package com.aqanetics.agent.testng.dto;
+package com.aqanetics.agent.core.dto;
 
 import java.time.Instant;
 import java.util.List;
