+++ src/main/java/com/aqanetics/agent/core/dto/OrganizationDto.java
@@ -1,4 +1,4 @@
-package com.aqanetics.agent.testng.dto;
+package com.aqanetics.agent.core.dto;
 
 public record OrganizationDto(String name) {
