+++ src/main/java/com/aqanetics/agent/core/dto/ParameterDto.java
@@ -1,3 +1,3 @@
-package com.aqanetics.agent.testng.dto;
+package com.aqanetics.agent.core.dto;
 
 public record ParameterDto(String type, Long valueNumeric, String valueText) {}
