+++ src/main/java/com/aqanetics/agent/core/dto/TestExecutionLogDto.java
@@ -1,4 +1,4 @@
-package com.aqanetics.agent.testng.dto;
+package com.aqanetics.agent.core.dto;
 
 import java.time.Instant;
