+++ src/main/java/com/aqanetics/agent/utils/AQATraceIgnore.java
@@ -1,4 +1,4 @@
-package com.aqanetics.utils;
+package com.aqanetics.agent.utils;
 
 import static java.lang.annotation.ElementType.METHOD;
 
@@ -8,11 +8,11 @@ import java.lang.annotation.Target;
 
 /**
  * Indicates that a specific test or configuration method should be ignored or skipped from being
- * registered in the AQA trace reporting software.
+ * registered in the AQA Trace reporting software.
  *
  * <p>When applied to a method (either a test method annotated with {@code @Test} or a configuration
  * method like {@code @BeforeMethod}, {@code @AfterMethod}, etc.), it specifically indicates that
- * this method should be ignored for AQA trace reporting.
+ * this method should be ignored for AQA Trace reporting.
  *
  * <p>The presence of this annotation, with its default value of {@code true}, is sufficient to
  * signal that the method should be ignored.
@@ -23,7 +23,7 @@ import java.lang.annotation.Target;
 public @interface AQATraceIgnore {
 
   /**
-   * Indicates whether the annotated method should be ignored for AQA trace reporting. The default
+   * Indicates whether the annotated method should be ignored for AQA Trace reporting. The default
    * value is {@code true}.
    *
    * @return {@code true} if the method should be ignored, {@code false} otherwise.
